requirements.txt
asgiref
Django
djangorestframework
djangorestframework-simplejwt
mysqlclient
Pillow
psycopg2-binary
PyJWT
python-dotenv
pytz
sqlparse
